+++ AWSPinpoint/AWSPinpointEndpointProfile.m
@@ -122,11 +122,11 @@ NSString *DEBUG_CHANNEL_TYPE = @"APNS_SANDBOX";
 }
 
 - (void) setEndpointOptOut:(BOOL) applicationLevelOptOut {
-    BOOL isNotificationEnabled = [AWSPinpointNotificationManager isNotificationEnabled];
-    BOOL isOptedOutForRemoteNotifications = !(isNotificationEnabled && [_address length]);
+    BOOL isUsingPinpointForNotifications = [AWSPinpointNotificationManager isNotificationEnabled] && [self.address length];
+    BOOL isOptedOutForNotifications = !isUsingPinpointForNotifications;
 
     @synchronized (self) {
-        self->_optOut = (applicationLevelOptOut || isOptedOutForRemoteNotifications)? @"ALL": @"NONE";
+        self->_optOut = (applicationLevelOptOut || isOptedOutForNotifications)? @"ALL": @"NONE";
     }
 }