+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # AWS Mobile SDK for iOS CHANGELOG
 
+## Unreleased
+
+### Bug Fixes
+- **Amazon Pinpoint**
+  - Fix Missing address issue related to apps which have enabled push notifications, is using the pinpoint SDK, but is not registering the token with the endpoint [PR: #2455](https://github.com/aws-amplify/aws-sdk-ios/pull/2455)
+
 ## 2.13.2
 
 ### Misc. Updates